+++ src/ProgressOnderwijsUtils/RequiredFields/NonNullableFieldVerifier0.cs
@@ -1,6 +1,4 @@
-using System.Linq.Expressions;
-
-namespace ProgressOnderwijsUtils.Tests;
+namespace ProgressOnderwijsUtils.RequiredFields;
 
 public static class NonNullableFieldVerifier0
 {
