+++ test/ProgressOnderwijsUtils.Tests/NonNullableNullabilityCheck.cs
@@ -1,3 +1,5 @@
+using ProgressOnderwijsUtils.RequiredFields;
+
 #pragma warning disable CS8625
 
 namespace ProgressOnderwijsUtils.Tests;
